enhancements_backend/noise/requirements.txt
torch
torchvision
librosa
pillow
numpy
fastapi
aiortc
uvicorn
python-socketio